agents/aws_strands/00-first-agent-strands/my_agent/requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0